django_api/requirements.txt
Django~=3.1.2
djangorestframework~=3.12.1
django-filter~=2.4.0
pytest~=6.1.2
model_bakery~=1.2.0